refactor(ship): move the shared HTTP fixtures into conftest so both suites discover them
